Separate auth in its own file
