+++ site/gatsby-site/src/server/local.ts
@@ -1,55 +1,11 @@
 
 import { makeExecutableSchema } from '@graphql-tools/schema';
-import { MapperKind, getDirective, mapSchema } from '@graphql-tools/utils';
-import { GraphQLSchema, defaultFieldResolver } from 'graphql';
 import { Resolvers } from './generated/graphql';
 import { LongResolver, ObjectIDResolver } from 'graphql-scalars';
-
 import localTypeDefs from './localTypeDefs';
 
 import { resolvers as quickAddsResolvers, typeDefs as quickAddsTypeDefs } from './fields/quickadds';
-
-
-function authDirectiveTransformer(schema: GraphQLSchema, directiveName: string) {
-
-    const typeDirectiveArgumentMaps: Record<string, any> = {}
-
-    return mapSchema(schema, {
-        [MapperKind.TYPE]: type => {
-            const authDirective = getDirective(schema, type, directiveName)?.[0]
-            if (authDirective) {
-                typeDirectiveArgumentMaps[type.name] = authDirective
-            }
-            return undefined
-        },
-        [MapperKind.OBJECT_FIELD]: (fieldConfig, _fieldName, typeName) => {
-            const authDirective =
-                getDirective(schema, fieldConfig, directiveName)?.[0] ??
-                typeDirectiveArgumentMaps[typeName]
-            if (authDirective) {
-                const { requires } = authDirective
-                if (requires) {
-
-                    const { resolve = defaultFieldResolver } = fieldConfig;
-
-                    fieldConfig.resolve = function (source, args, context, info) {
-
-                        const { user } = context;
-
-                        if (!user || !user.roles || !user.roles.includes(requires)) {
-
-                            throw new Error('not authorized')
-                        }
-
-                        return resolve(source, args, context, info)
-                    }
-                    return fieldConfig
-                }
-            }
-        }
-    })
-}
-
+import { transformer as authDirectiveTransformer, typeDefs as authTypeDefs } from './directives/auth';
 
 export const getSchema = async () => {
 
@@ -65,7 +21,7 @@ export const getSchema = async () => {
     };
 
 
-    const schema = makeExecutableSchema({ typeDefs: [localTypeDefs, quickAddsTypeDefs], resolvers });
+    const schema = makeExecutableSchema({ typeDefs: [localTypeDefs, authTypeDefs, quickAddsTypeDefs], resolvers });
 
     const schemaWithAuth = authDirectiveTransformer(schema, 'auth');
