+++ site/gatsby-site/src/server/localTypeDefs.ts
@@ -1,12 +1,6 @@
 import gql from "graphql-tag";
 
 export default gql`
-directive @auth(requires: Role = admin) on OBJECT | FIELD_DEFINITION
-
-enum Role {
-  admin
-  subscriber
-}
 
 scalar Long
 scalar ObjectId
